Adds ability test queue-csv and queue-bamboohr using --dry option
- adds limit to queue-csv

+++ src/queue/inserter.js
@@ -12,14 +12,26 @@ module.exports = (config, log = {}) => {
     }
   }
 
+  log.processedCount = 0
   log.errorCount = 0
   log.cachedHitCount = 0
+  log.cachedCount = 0
   log.insertedCount = 0
+  log.sample = null
+
+  if( config.dryRun ){
+    log.willFilter = []
+    log.cacheHits = []
+    log.willCache = []
+    log.willInsert = []
+  }
 
   const filter = (typeof config.filter === 'function') ? config.filter : record => record
 
   return async data => {
     let reduced = {}
+
+    log.processedCount++
 
     mapper.forEach(m => {
       m(data, reduced)
@@ -40,6 +52,9 @@ module.exports = (config, log = {}) => {
     log.sample = reduced
 
     if( !filter(reduced) ){
+      if( config.dryRun ){
+        log.willFilter.push(reduced)
+      }
       return false
     }
     
@@ -56,17 +71,33 @@ module.exports = (config, log = {}) => {
 
       if (cached === reducedJSON) {
         log.cachedHitCount++
+        if( config.dryRun ){
+          log.cacheHits.push(reduced)
+        }
         return
       }
+      
+      if( config.dryRun ){
+        log.willCache.push( reduced )
+      }
+      else{
+        log.cachedCount++
+        fs.writeFileSync(_cachePath, reducedJSON)
+      }
 
-      fs.writeFileSync(_cachePath, reducedJSON)
     }
 
     let path = `${qPath}/${id}.json`
-    fs.writeFileSync(path, reducedJSON)
 
-    
+    if( config.dryRun ){
+      log.willInsert.push(reduced)
+    }
+    else{
+      fs.writeFileSync(path, reducedJSON)
+    }
+
     log.insertedCount++
+    
   }
 
 }
